fix(ios): issue 621 - getLocation not resolving
It seems this issue is only happening if flutter is
already requesting location updates when the call to
getLocation is made.

Potential fix:
In the case where location updates are already running,
and a valid location is already present, return the
latest location to the caller when getLocation is called.